refactor: ErrorMessage
