+++ pokemon_app/lib/features/messages/error_message.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
 
 class ErrorMessage extends StatelessWidget {
+  final VoidCallback? onPressedTryAgain;
+
   const ErrorMessage({
     super.key,
+    required this.onPressedTryAgain,
   });
 
   @override
@@ -12,17 +15,28 @@ class ErrorMessage extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
-          Text(
-            'Failed to get data.',
-            style: Theme.of(context).textTheme.bodyMedium,
+          Center(
+            child: Column(
+              children: [
+                Text(
+                  'Failed to get data.',
+                  style: Theme.of(context).textTheme.bodyMedium,
+                ),
+                Text(
+                  'Please, check your internet connection.',
+                  style: Theme.of(context).textTheme.bodySmall,
+                  textAlign: TextAlign.center,
+                ),
+                const SizedBox(
+                  height: 30,
+                ),
+              ],
+            ),
           ),
-          Text(
-            'Please, check your internet connection.',
-            style: Theme.of(context).textTheme.bodySmall,
-            textAlign: TextAlign.center,
-          ),
-          const SizedBox(
-            height: 30,
+          TextButton(
+            onPressed: onPressedTryAgain,
+            child:
+                Text('Try again', style: Theme.of(context).textTheme.bodySmall),
           ),
         ],
       ),
