+++ pokemon_app/lib/features/pokemons/view/pokemons_screen.dart
@@ -12,14 +12,16 @@ class _PokemonsState extends State<Pokemons> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: Text('Pokemons List',
-              style: Theme.of(context).textTheme.headlineMedium),
-        ),
-        body: const Center(
-            child: SizedBox(
+      appBar: AppBar(
+        title: Text('Pokemons List',
+            style: Theme.of(context).textTheme.headlineMedium),
+      ),
+      body: const Center(
+        child: SizedBox(
           height: 500,
           child: PokemonsList(),
-        )));
+        ),
+      ),
+    );
   }
 }
