+++ pokemon_app/lib/repositories/pokemons_rep_export.dart
@@ -1,3 +1,4 @@
-export 'abstract_pokemons_list_rep.dart';
+export 'abstract_pokemons_rep.dart';
 export 'pokemons_list_rep.dart';
+export 'pokemons_info_rep.dart';
 export 'models/models_export.dart';
